+++ internal/client/client.go
@@ -168,6 +168,8 @@ func (watcher *Watcher) waitForDeployment(id, appName, version string) error {
 			return fmt.Errorf("Application %s does not exist.\n%s", appName, taskInfo.StatusReason)
 		case models.StatusArgoCDUnavailableMessage:
 			return fmt.Errorf("ArgoCD is unavailable. Please investigate.\n%s", taskInfo.StatusReason)
+		case models.StatusAborted:
+			return fmt.Errorf("The deployment was aborted before its outcome could be confirmed. See the reason below.\n%s", taskInfo.StatusReason)
 		case models.StatusDeployedMessage:
 			log.Printf("The deployment of %s version is done.", version)
 			return nil